Break out Leaflet markers.
